ml/requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
scipy>=1.11.0
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
huggingface-hub>=0.19.0
httpx>=0.25.0  # Async HTTP client for MastraAI backend

# RAG dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Use faiss-gpu for CUDA support
langchain>=0.1.0
tiktoken>=0.5.2

# Multimodal dependencies
Pillow>=10.1.0
pytesseract>=0.3.10  # OCR for screenshots
opencv-python>=4.8.1
timm>=0.9.12  # Vision models

# Training & monitoring
tensorboard>=2.15.0
wandb>=0.16.0
